fix issue 180: assert when double coordinate is +-Infinite
X-Committed-on-Branch: maintenance

+++ ui/org.omnetpp.scave/src/org/omnetpp/scave/charting/plotter/LinesVectorPlotter.java
@@ -94,7 +94,7 @@ public class LinesVectorPlotter extends VectorPlotter {
 				prevX = x;
 			}
 			else {
-				prevX = Integer.MIN_VALUE; // invalidate minX/maxX
+				prevX = Long.MIN_VALUE; // invalidate minX/maxX
 			}
 			prevY = y;
